CrossRefApiCitationsScrapper added and ScholarCitationsScrapper disable to prevent IP banning

+++ src/main/java/tfg/urjc/mydoiinfo/scrappers/citationsScrappers/CitationsScrapper.java
@@ -4,5 +4,5 @@ import tfg.urjc.mydoiinfo.domain.entities.Article;
 import tfg.urjc.mydoiinfo.scrappers.Scrapper;
 
 public abstract class CitationsScrapper extends Scrapper {
-    public abstract Integer getCitationsFromArticle(Article article);
+    public abstract Long getCitationsFromArticle(Article article);
 }
